% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d.R
\name{inla.mesh.fem}
\alias{inla.mesh.fem}
\alias{inla.mesh.fem.inla_mesh_3d}
\alias{inla.mesh.fem.default}
\alias{inla.mesh3d.fem}
\alias{inla.mesh3d.volumes}
\title{FUNCTION_TITLE}
\usage{
inla.mesh.fem(mesh, ...)

\method{inla.mesh.fem}{inla_mesh_3d}(mesh, ...)

\method{inla.mesh.fem}{default}(mesh, ...)

inla.mesh3d.fem(mesh, ...)

inla.mesh3d.volumes(mesh, ...)
}
\arguments{
\item{mesh}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
