% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d.R
\name{inla.mesh3d}
\alias{inla.mesh3d}
\alias{inla_mesh_3d}
\title{FUNCTION_TITLE}
\usage{
inla.mesh3d(loc, tv)
}
\arguments{
\item{loc}{PARAM_DESCRIPTION}

\item{tv}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
