% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d.R
\name{inla.spde2.matern}
\alias{inla.spde2.matern}
\alias{inla.spde2.matern.inla_mesh_3d}
\alias{inla.spde2.matern.default}
\alias{inla.spde2.matern3d}
\title{FUNCTION_TITLE}
\usage{
inla.spde2.matern(mesh, ...)

\method{inla.spde2.matern}{inla_mesh_3d}(mesh, ...)

\method{inla.spde2.matern}{default}(mesh, ...)

inla.spde2.matern3d(
  mesh,
  param = NULL,
  constr = FALSE,
  extraconstr.int = NULL,
  extraconstr = NULL,
  fractional.method = c("parsimonious", "null"),
  n.iid.group = 1,
  ...
)
}
\arguments{
\item{mesh}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{param}{PARAM_DESCRIPTION, Default: NULL}

\item{constr}{PARAM_DESCRIPTION, Default: FALSE}

\item{extraconstr.int}{PARAM_DESCRIPTION, Default: NULL}

\item{extraconstr}{PARAM_DESCRIPTION, Default: NULL}

\item{fractional.method}{PARAM_DESCRIPTION, Default: c("parsimonious", "null")}

\item{n.iid.group}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link{param2.matern3d}},\code{\link{param2.matern_calc}}
}
\keyword{Internal}
