% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d.R
\name{param2.matern}
\alias{param2.matern}
\alias{param2.matern.inla_mesh_3d}
\alias{param2.matern.inla.mesh}
\alias{param2.matern.inla.mesh.1d}
\alias{param2.matern_calc}
\title{FUNCTION_TITLE}
\usage{
param2.matern(mesh, ...)

\method{param2.matern}{inla_mesh_3d}(mesh, ...)

\method{param2.matern}{inla.mesh}(mesh, ...)

\method{param2.matern}{inla.mesh.1d}(mesh, ...)

param2.matern_calc(
  dim = NULL,
  dof = NULL,
  alpha = 2,
  prior_range = NULL,
  prior_sigma = NULL,
  B_range = matrix(c(0, 1, 0), 1, 3),
  B_sigma = matrix(c(0, 0, 1), 1, 3),
  prior_theta = NULL,
  ...
)
}
\arguments{
\item{mesh}{PARAM_DESCRIPTION}

\item{\dots}{PARAM_DESCRIPTION}

\item{dim}{PARAM_DESCRIPTION, Default: NULL}

\item{dof}{PARAM_DESCRIPTION, Default: NULL}

\item{alpha}{PARAM_DESCRIPTION, Default: 2}

\item{prior_range}{\code{list(median, factor)},
Default: \code{factor=5}}

\item{prior_sigma}{\code{list(median, factor)},
Default: \code{factor=5}}

\item{B_range}{PARAM_DESCRIPTION, Default: \code{matrix(c(0, 1, 0), 1, 3)}}

\item{B_sigma}{PARAM_DESCRIPTION, Default: \code{matrix(c(0, 0, 1), 1, 3)}}

\item{prior_theta}{If not \code{NULL}, \code{prior_range} and \code{prior_sigma} will only
be used to set the offset columns of \code{B_range} and \code{B_sigma}, and not to set
the prior distribution for the \code{theta} parameters.
A list with elements \code{mean} and \code{prec}.
Default: \code{NULL}}

\item{...}{
  Arguments passed on to \code{\link[=param2.matern_calc]{param2.matern_calc}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

The priors for \code{range} (the spatial correlation range parameter)
and \code{sigma} (the field standard deviation parameter) are constructed so that
the prior median is given by \code{median}, and
\deqn{P(median/factor < param < median*factor) = 0.98}
which means that the std.dev. of \code{log(param)} is \code{log(factor)/stats::qnorm(0.99)}.

Remark: In the old \code{param2.matern.orig} function, the internal parameter
scale used a default prior precision \eqn{0.1}. This corresponds to
\code{factor=exp(stats::qnorm((1+0.98)/2)/sqrt(0.1))}, that equals \eqn{1566.435}.
This large value has lead to numerical and other problems for many models.
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
