% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d.R
\name{inla.mesh3d.bary}
\alias{inla.mesh3d.bary}
\title{FUNCTION_TITLE}
\usage{
inla.mesh3d.bary(mesh, loc, divide_along = 1, t_subset = NULL)
}
\arguments{
\item{mesh}{PARAM_DESCRIPTION}

\item{loc}{PARAM_DESCRIPTION}

\item{divide_along}{PARAM_DESCRIPTION}

\item{t_subset}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
