% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d.R
\name{inla.mesh3d.make.A}
\alias{inla.mesh3d.make.A}
\title{FUNCTION_TITLE}
\usage{
inla.mesh3d.make.A(mesh, loc, ...)
}
\arguments{
\item{mesh}{PARAM_DESCRIPTION}

\item{loc}{PARAM_DESCRIPTION}

\item{...}{Unused}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[Matrix]{sparseMatrix}}
}
