% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d.R
\name{inla.spde.make.A}
\alias{inla.spde.make.A}
\alias{fm_basis.inla_mesh_3d}
\alias{fm_evaluator.inla_mesh_3d}
\title{FUNCTION_TITLE}
\usage{
\method{fm_basis}{inla_mesh_3d}(x, ...)

\method{fm_evaluator}{inla_mesh_3d}(mesh, loc = NULL, ...)
}
\arguments{
\item{x}{3D mesh}

\item{...}{Passed on to \code{\link[fmesher:fm_basis]{fmesher::fm_basis()}}}

\item{mesh}{An \code{inla_mesh_3d} mesh}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
