% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d.R
\name{inla.spde.make.A}
\alias{inla.spde.make.A}
\alias{inla.spde.make.A.inla_mesh_3d}
\alias{inla.spde.make.A.default}
\title{FUNCTION_TITLE}
\usage{
inla.spde.make.A(mesh, ...)

\method{inla.spde.make.A}{inla_mesh_3d}(mesh, ...)

\method{inla.spde.make.A}{default}(mesh, ...)
}
\arguments{
\item{mesh}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
