% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d.R
\name{param2.matern}
\alias{param2.matern}
\alias{param2.matern.inla_mesh_3d}
\alias{param2.matern.inla.mesh}
\alias{param2.matern.inla.mesh.1d}
\title{FUNCTION_TITLE}
\usage{
param2.matern(mesh, ...)

\method{param2.matern}{inla_mesh_3d}(mesh, ...)

\method{param2.matern}{inla.mesh}(mesh, ...)

\method{param2.matern}{inla.mesh.1d}(mesh, ...)
}
\arguments{
\item{mesh}{PARAM_DESCRIPTION}

\item{...}{
  Arguments passed on to \code{\link[=param2.matern_calc]{param2.matern_calc}}
  \describe{
    \item{\code{alpha}}{PARAM_DESCRIPTION, Default: 2}
    \item{\code{prior_range}}{\code{list(median, factor)},
Default: \code{factor=5}}
    \item{\code{prior_sigma}}{\code{list(median, factor)},
Default: \code{factor=5}}
    \item{\code{B_range}}{PARAM_DESCRIPTION, Default: \code{matrix(c(0, 1, 0), 1, 3)}}
    \item{\code{B_sigma}}{PARAM_DESCRIPTION, Default: \code{matrix(c(0, 0, 1), 1, 3)}}
    \item{\code{prior_theta}}{If not \code{NULL}, \code{prior_range} and \code{prior_sigma} will only
be used to set the offset columns of \code{B_range} and \code{B_sigma}, and not to set
the prior distribution for the \code{theta} parameters.
A list with elements \code{mean} and \code{prec}.
Default: \code{NULL}}
  }}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
