% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d.R
\name{row_cross_product}
\alias{row_cross_product}
\title{FUNCTION_TITLE}
\usage{
row_cross_product(e1, e2)
}
\arguments{
\item{e1}{PARAM_DESCRIPTION}

\item{e2}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
